Acces Control and stuff like that

+++ app/Http/Controllers/PostsController.php
@@ -10,6 +10,14 @@ use App\Models\Post ;
 //use DB ;
 class PostsController extends Controller
 {
+ //acces control
+ public function __construct()
+ {
+     $this->middleware('auth', ['except' => ['index', 'show']]);
+     //
+ }
+
+
     /**
      * Display a listing of the resource.
      *
@@ -55,6 +63,7 @@ class PostsController extends Controller
         $post = new Post ; 
         $post->title =$request->input('title');
         $post->body = $request->input('body') ; 
+        $post->user_id= auth()->user()->id;
         $post->save();
         return redirect('/posts')->with('success','post created');
     }
@@ -81,7 +90,11 @@ class PostsController extends Controller
     public function edit($id)
     {
         //
+      
         $post = Post::find($id);
+         if (auth()->user()->id!==$post->user_id){
+              return redirect ('/posts')->with('error','SORRY ! You are unautherized to edit this post ');
+          }
         return view ('posts.edit' ,['post'=>$post]);
 
     }
@@ -130,6 +143,10 @@ class PostsController extends Controller
     {
         $post=Post::find($id);
         $title=$post->title ;
+      
+        if (auth()->user()->id!==$post->user_id){
+            return redirect ('/posts')->with('error','SORRY ! You are unautherized to delete this post ');
+        }  
         $post->delete();
         return redirect('/posts' )->with('success' , $title.'  deleted');
         //
